pyproject.toml
[project]
name = "phmlondon"
version = "0.1.0"
description = "Helper functions for population health data science on OneLondon Snowflake"
authors = [
    { name = "Joe Zhang", email = "joe.zhang@gstt.nhs.uk" },
    { name = "Lawrence Adams", email = "lawrence.adams@gstt.nhs.uk" },
    { name = "Isobel Weinberg", email = "isobel.weinberg@kcl.ac.uk" },
    { name = "Dan Stein", email = "dan.stein@gstt.nhs.uk"},
    { name = "Jasjot Saund", email = "jasjot.saund@gstt.nhs.uk"}
]
readme = "README.md"
requires-python = ">=3.11.2, <3.12" #snowpark requires python 3.11
dependencies = [
    "snowflake-connector-python==3.12.2",
    "snowflake-snowpark-python==1.22.1",
    "pandas==2.2.2",
    "polars==1.24.0",
    "tqdm",
    "PyYAML",
    "pyconceptlibraryclient @ git+https://github.com/SwanseaUniversityMedical/pyconceptlibraryclient.git@v1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/phmlondon", "phenolab"]

[tool.ruff]
 line-length = 120

 target-version = "py311"

 [tool.ruff.lint]
 select = [
     "E",   # pycodestyle: bad formatting
     "W",   # pycodestyle: warnings
     "F",   # pyflakes: common errors
     "I",   # isort: handles badly setup imports, and orders
     "B",   # flake8-bugbear: common bugs/footguns
     "C90", # mccabe cyclomatic complexity: for badly nested conditional statements that makes logic hard to follow
 ]